split rrdcontext.c to multiple files
