+++ configure.ac
@@ -1711,6 +1711,7 @@ AC_CONFIG_FILES([
     collectors/perf.plugin/Makefile
     daemon/Makefile
     database/Makefile
+    database/contexts/Makefile
     database/engine/Makefile
     database/ram/Makefile
     database/sqlite/Makefile
